refimpact: API wrapper for the UK REF 2014 Impact Case Studies Database

This package provides wrapper functions around the UK Research Excellence Framework 2014 Impact Case Studies Database API (http://impact.ref.ac.uk/). The database contains relevant publication and research metadata about each case study as well as several paragraphs of text from the case study submissions. Case studies in the database are licenced under a CC-BY 4.0 licence (http://creativecommons.org/licenses/by/4.0/legalcode).

Details

To get started, see ref_get and read the vignette using vignette('refimpact').